// src/rate/type.ts
export type DailyRate = {
  disclaimer: string;
  date: Date;
  timestamp: number;
  base: "RUB";
  rates: {
    AUD: number;
    AZN: number;
    GBP: number;
    AMD: number;
    BYN: number;
    BGN: number;
    BRL: number;
    HUF: number;
    VND: number;
    HKD: number;
    GEL: number;
    DKK: number;
    AED: number;
    USD: number;
    EUR: number;
    EGP: number;
    INR: number;
    IDR: number;
    KZT: number;
    CAD: number;
    QAR: number;
    KGS: number;
    CNY: number;
    MDL: number;
    NZD: number;
    NOK: number;
    PLN: number;
    RON: number;
    XDR: number;
    SGD: number;
    TJS: number;
    THB: number;
    TRY: number;
    TMT: number;
    UZS: number;
    UAH: number;
    CZK: number;
    SEK: number;
    CHF: number;
    RSD: number;
    ZAR: number;
    KRW: number;
    JPY: number;
  };
};